# docs/swagger.yaml
definitions:
  api.ErrorResponse:
    properties:
      message:
        example: Error message
        type: string
      status:
        example: error
        type: string
    type: object
  api.HealthResponse:
    properties:
      apiVersion:
        example: v1
        type: string
      buildVersion:
        example: 1.2.3
        type: string
      httpServer:
        example: true
        type: boolean
      listenerIsRunning:
        example: true
        type: boolean
      message:
        example: pong
        type: string
      status:
        example: ok
        type: string
    type: object
  api.MessageResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      fail_reason:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      response_message_id:
        type: string
      retry_count:
        type: integer
      sent_at:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  api.NewMessageRequest:
    properties:
      content:
        example: Hello, this is a test message.
        type: string
      phone_number:
        example: +905xxxxxxxx
        type: string
    type: object
  api.NewMessageResponse:
    properties:
      message:
        example: Successfully added new message
        type: string
      status:
        example: ok
        type: string
    type: object
  api.PaginatedMessagesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/api.MessageResponse'
        type: array
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      returned_count:
        example: 20
        type: integer
      status:
        example: ok
        type: string
    type: object
  api.SimpleResponse:
    properties:
      isRunning:
        example: true
        type: boolean
      message:
        example: Started listener
        type: string
      status:
        example: ok
        type: string
    type: object
info:
  contact: {}
paths:
  /api/messages:
    post:
      consumes:
      - application/json
      description: Adds a new message to the queue for sending
      parameters:
      - description: Message to add
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/api.NewMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Message added to queue
          schema:
            $ref: '#/definitions/api.NewMessageResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Failed to add
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Add new message
      tags:
      - messages
  /api/messages/sent:
    get:
      description: Returns a paginated list of sent messages
      parameters:
      - description: Number of messages to return (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PaginatedMessagesResponse'
        "500":
          description: Failed to retrieve messages
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List sent messages
      tags:
      - messages
  /api/start-listener:
    get:
      description: Starts the automatic message sending loop
      produces:
      - application/json
      responses:
        "200":
          description: started
          schema:
            $ref: '#/definitions/api.SimpleResponse'
        "409":
          description: already running
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Start Listener
      tags:
      - listener
  /api/stop-listener:
    get:
      description: Stops the automatic message sending loop
      produces:
      - application/json
      responses:
        "200":
          description: stopped
          schema:
            $ref: '#/definitions/api.SimpleResponse'
        "409":
          description: not running
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Stop Listener
      tags:
      - listener
  /ping:
    get:
      description: Returns simple pong response with server and listener status. You
        can also learn the API version.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HealthResponse'
      summary: Health check
      tags:
      - health
swagger: "2.0"